backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
scikit-learn==1.4.0
werkzeug==3.0.1
gunicorn==21.2.0
